+++ tf-modules/ingest/variables.tf
@@ -64,6 +64,12 @@ variable "default_s3_multipart_chunksize_mb" {
   default = 256
 }
 
+variable "enable_mismatch" {
+  description = "optional variable to be used in message_consumer lambdas for disabling rule/message provider mismatches"
+  type = string
+  default = 'true'
+}
+
 variable "distribution_url" {
   type = string
 }